8 indicators STEP data
Script to prepare 8 consistent indicators to measure deprivation in STEP data. This version assumes that BMI is the accurate standard to measure under nutrition for people 18 years and older and assumes this as the standard for nutrition. It also assumes that lower primary is an indicator of deprivation. Other data is not kept in the summary created at the end of the file.